# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignIn": {
        "additionalProperties": false,
        "properties": {
          "assignee_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "User id, or null to unassign",
            "title": "Assignee Id"
          }
        },
        "required": [
          "assignee_id"
        ],
        "title": "AssignIn",
        "type": "object"
      },
...
      "AssigneesIn": {
        "properties": {
          "user_ids": {
            "description": "In order; the first leads",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 10,
            "title": "User Ids",
            "type": "array"
          }
        },
        "required": [
          "user_ids"
        ],
        "title": "AssigneesIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "total"
        ],
        "title": "IncidentPage",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "Status": {
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "closed"
        ],
        "title": "Status",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-assisted incident management. All business routes live under /api/v1.",
    "title": "Tasky API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/Status"
              },
              "title": "Status",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/Priority"
              },
              "title": "Priority",
              "type": "array"
            }
          },
          {
            "description": "User id, `me` or `none`; repeatable",
            "in": "query",
            "name": "assignee",
            "required": false,
            "schema": {
              "default": [],
              "description": "User id, `me` or `none`; repeatable",
              "items": {
                "type": "string"
              },
              "title": "Assignee",
              "type": "array"
            }
          },
          {
            "description": "User id or `me`",
            "in": "query",
            "name": "reporter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "User id or `me`",
              "title": "Reporter"
            }
          },
          {
            "description": "Full-text + fuzzy title search; `INC-142` too",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text + fuzzy title search; `INC-142` too",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sla",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "breached",
                    "at_risk",
                    "on_track"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sla"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "description": "All given tags must match",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "default": [],
              "description": "All given tags must match",
              "items": {
                "type": "string"
              },
              "title": "Tag",
              "type": "array"
            }
          },
          {
            "description": "Admins: list the recycle bin",
            "in": "query",
            "name": "deleted",
            "required": false,
            "schema": {
              "default": false,
              "description": "Admins: list the recycle bin",
              "title": "Deleted",
              "type": "boolean"
            }
          },
          {
            "description": "Only incidents I watch",
            "in": "query",
            "name": "watching",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only incidents I watch",
              "title": "Watching",
              "type": "boolean"
            }
          },
          {
            "description": "Comma-separated, `-` = descending. Allowed: created_at, updated_at, priority, status, number, resolution_due_at, relevance (needs q). Default: -created_at, or relevance when q is given.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated, `-` = descending. Allowed: created_at, updated_at, priority, status, number, resolution_due_at, relevance (needs q). Default: -created_at, or relevance when q is given.",
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List and search incidents",
        "tags": [
          "incidents"
        ]
      },
      "post": {
        "description": "The current user is the reporter. SLA due times are set from the priority.",
        "operationId": "create_incident_api_v1_incidents_post",
        "parameters": [
          {
            "description": "Retries with the same key return the original incident instead of creating a duplicate.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Retries with the same key return the original incident instead of creating a duplicate.",
              "title": "Idempotency-Key"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create an incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{ident}": {
      "delete": {
        "operationId": "delete_incident_api_v1_incidents__ident__delete",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Soft-delete (admin)",
        "tags": [
          "incidents"
        ]
      },
      "get": {
        "operationId": "get_incident_api_v1_incidents__ident__get",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get an incident by id or INC-key",
        "tags": [
          "incidents"
        ]
      },
      "patch": {
        "operationId": "update_incident_api_v1_incidents__ident__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update fields (audited)",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/api/v1/incidents/{ident}/assign": {
      "post": {
        "description": "Replaces every assignee with this one person (null: nobody). Use PUT /assignees to assign several people.",
        "operationId": "assign_incident_api_v1_incidents__ident__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign to one person, or unassign",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/v1/incidents/{ident}/assignees": {
      "put": {
        "description": "The full list, in order (the first is the lead). People newly on the list are emailed; an empty list unassigns everyone (not while in progress).",
        "operationId": "set_assignees_api_v1_incidents__ident__assignees_put",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssigneesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set everyone assigned",
        "tags": [
          "incidents"
        ]
      }
    },
...
    "/api/v1/incidents/{ident}/events": {
      "get": {
        "operationId": "list_events_api_v1_incidents__ident__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "ident",
            "required": true,
            "schema": {
              "title": "Ident",
              "type": "string"
            }
          },
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response List Events Api V1 Incidents  Ident  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit timeline",
        "tags": [
          "incidents"
        ]
      }
    },
...
  }
}